[#121] chore: Gptclient 인프라 패키지로 변경
