+++ src/test/kotlin/chat/teco/tecochat/domain/keyword/KeywordExtractorTest.kt
@@ -3,6 +3,7 @@ package chat.teco.tecochat.domain.keyword
 import chat.teco.tecochat.createChat
 import chat.teco.tecochat.createQuestionAndAnswer
 import chat.teco.tecochat.domain.chat.Answer
+import chat.teco.tecochat.infra.gpt.GptClient
 import io.kotest.assertions.extracting
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.collections.shouldContainExactly
